feat: update room data

+++ hotel_api/hotel_api/ApiDto/Room/RoomRequestDto.cs
@@ -15,6 +15,6 @@ public class RoomRequestDto
 
     public List<ImageRequestDto>? images { get; set; }
     public string? location { get; set; }
-    public double?  latitude { get; set; }
-    public double? longitude { get; set; }
+    public decimal?  latitude { get; set; }
+    public decimal? longitude { get; set; }
 }
